refactor(tests): consolidate Rust integration tests into a single binary
Cargo compiles one test executable per top-level file under `tests/`, and
each one statically links the full `taomni_lib` with debug info (hundreds of
MB apiece). With ~14 separate integration files this multiplied into several
GB of `target/debug` and repeated link time on every `cargo test`.

Collapse all integration cases into one crate so the heavy lib is linked
exactly once.

Structural changes
- Move every former `tests/<case>.rs` into `tests/integration/<case>.rs` and
  register them as modules of a single crate root `tests/integration/main.rs`.
  Cargo now builds one `--test integration` binary instead of ~14.
- Move the shared mock helper to `tests/integration/support/` and switch the
  four consumers (concurrent_three_workloads, fim_latency, router_routing,
  voice_intent_latency) from per-file `mod support; use support::...` to the
  crate-shared `use crate::support::...`. No test logic changed.
- Keep `tests/support/acp_fake_agent.rs` in place: it is the `[[bin]]
  acp-fake-agent` fixture (declared in Cargo.toml), not a test target, so it
  must stay out of the integration crate.

Disk / build-profile tuning (Cargo.toml)
- profile.dev / profile.test: debug = 1 (line tables only) — keeps usable
  backtraces while shrinking rlibs, the main binary, incremental caches, and
  test executables.
- profile.dev.package."*".debug = false — drop debug info for dependency
  crates entirely; the local package is unaffected.

Docs
- AGENTS.md: document the new command matrix — `cargo test --lib` (fast,
  no integration binary), `cargo test --test integration` (unified suite),
  `cargo test` (full).
- tests/README.md: describe the single-binary layout, day-to-day commands,
  and the disk/profile notes.

Merge-carried change
- tests/integration/router_vault_lock.rs picks up main's new required
  `AiConfig` field `chat_send_shortcut: "ctrl_enter".into()` (added by the
  chat send-shortcut feature). Without it the integration crate would not
  compile after merging main.

Verification
- `cargo test --test integration --no-run` links the unified binary cleanly
  (only pre-existing unused-import / dead-code warnings; no errors).
