+++ src-tauri/tests/integration/asr_llm_isolation.rs
