+++ src-tauri/tests/integration/cloud_only_smoke.rs
