+++ src-tauri/tests/integration/gpu_detect.rs
