+++ src-tauri/tests/integration/models_downloader.rs
