+++ src-tauri/tests/integration/router_vault_lock.rs
