+++ src-tauri/tests/integration/sockscap_win11_scenarios.rs
