+++ src-tauri/tests/integration/support/mock_provider.rs
