+++ src-tauri/tests/integration/three_source_probe.rs
